requirements.txt
discord.py
asyncio
aiohttp
dataIO
requests
bs4
